refactor CDN update tests to be parameterized for CDN service instance types
