+++ tests/integration/test_cdn_provision_common_checks.py
